requirements.txt
python-telegram-bot==20.7
openai>=1.40.0
ffmpeg-python==0.2.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx>=0.24.0

